+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # eppz.Persistence
 
+* 0.2.0
+
+	+ Using `System.IO.Compression`
+	+ Using `Stream.CopyTo()`
+	+ String extensions `Zip()` and `Unzip()` works fine
+	+ Gives the same results as other libraries (like `pako.js`)
+
 * 0.0.1
 
 	+ Initial commit